+++ src/main/java/org/example/calculation/Calculation.java
@@ -1,5 +1,5 @@
 package org.example.calculation;
 
 public interface Calculation {
-    int run(String input);
+    double run(String input);
 }
