+++ src/main/java/org/example/io/Output.java
@@ -3,6 +3,6 @@ package org.example.io;
 public interface Output {
 
     void printAction(String action);
-    void printCaculatedResult(int result);
+    void printCaculatedResult(double result);
     void printFindAll(String[] strings);
 }
